requirements.txt
chromadb
fastmcp
google-genai
httpx
markdown-it-py
mcp
psutil
pydantic
PyQt6
Pygments
python-dotenv
qasync
ruff
langchain
langchain-google-genai
langchain-community
nest_asyncio
fastapi
uvicorn
websockets
pymupdf
python-docx
pyperclip
pyautogui
Pillow
sqlalchemy
matplotlib
seaborn
pandas
networkx
scikit-learn
rank_bm25
sentence-transformers
huggingface-hub>=0.34.0,<1.0
pdfplumber
openpyxl
python-pptx
playwright
diskcache
